requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiometer==0.5.0
aiosignal==1.3.2
aiosqlite==0.20.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4>=4.12.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cohere==5.13.4
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
docstring-parser==0.16
docx==0.2.4
ebooklib>=0.18.0
et-xmlfile==2.0.0
eval-type-backport==0.2.2
executing==2.1.0
fastavro==1.10.0
filelock==3.16.1
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
groq==0.13.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gtts==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
lxml==5.3.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.2.5
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.19.1
nest-asyncio==1.6.0
numpy==2.2.1
openai==1.58.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow>=10.0.0
pip==24.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.4
pydantic-core==2.27.2
pydeck==0.9.1
pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pymupdf==1.25.3
pypdf2==3.0.1
pytesseract>=0.3.10
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
python-pptx==1.0.2
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
ratelimit==2.2.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.6.0
shapely==2.0.6
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlite3>=3.35.0
stack-data==0.6.3
streamlit>=1.30.0
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomli-w==1.1.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20241016
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.3.0
uv==0.5.11
vertexai==1.71.1
watchdog==6.0.0
wcwidth==0.2.13
whisper==1.1.10
xlsxwriter==3.2.0
yarl==1.18.3
